Added CAN_DELETE check to delete method of EntityHandler
Also added CAN_DELTE constant to PelagosEntityVoter

+++ src/Pelagos/Bundle/AppBundle/Handler/EntityHandler.php
@@ -142,6 +142,11 @@ class EntityHandler
      */
     public function delete(Entity $entity)
     {
+        if (!$this->authorizationChecker->isGranted(PelagosEntityVoter::CAN_DELETE, $entity)) {
+            throw new AccessDeniedException(
+                'You do not have sufficient privileges to delete this ' . $entity::FRIENDLY_NAME . '.'
+            );
+        }
         $this->entityManager->remove($entity);
         $this->entityManager->flush();
         return $entity;
